master

+++ pkg/apis/sensor/v1alpha1/types.go
@@ -194,7 +194,7 @@ type Subscription struct {
 // HTTPSubscription holds the context of the HTTP subscription of events for the sensor.
 type HTTPSubscription struct {
 	// Port on which sensor server should run.
-	Port int `json:"port" protobuf:"bytes,1,name=port"`
+	Port int32 `json:"port" protobuf:"bytes,1,name=port"`
 }
 
 // NATSSubscription holds the context of the NATS subscription of events for the sensor
@@ -415,7 +415,7 @@ type HTTPTrigger struct {
 	// Timeout refers to the HTTP request timeout in seconds.
 	// Default value is 60 seconds.
 	// +optional
-	Timeout int `json:"timeout,omitempty" protobuf:"bytes,6,opt,name=timeout"`
+	Timeout int64 `json:"timeout,omitempty" protobuf:"bytes,6,opt,name=timeout"`
 	// BasicAuth configuration for the http request.
 	// +optional
 	BasicAuth *BasicAuth `json:"basicAuth,omitempty" protobuf:"bytes,7,opt,name=basicAuth"`
@@ -479,14 +479,14 @@ type KafkaTrigger struct {
 	// More info at https://kafka.apache.org/documentation/#intro_topics
 	Topic string `json:"topic" protobuf:"bytes,2,name=topic"`
 	// Partition to write data to.
-	Partition int `json:"partition" protobuf:"bytes,3,name=partition"`
+	Partition int32 `json:"partition" protobuf:"bytes,3,name=partition"`
 	// Parameters is the list of parameters that is applied to resolved Kafka trigger object.
 	// +listType=triggerParameters
 	Parameters []TriggerParameter `json:"parameters,omitempty" protobuf:"bytes,4,rep,name=parameters"`
 	// RequiredAcks used in producer to tell the broker how many replica acknowledgements
 	// Defaults to 1 (Only wait for the leader to ack).
 	// +optional.
-	RequiredAcks int `json:"requiredAcks,omitempty" protobuf:"bytes,5,opt,name=requiredAcks"`
+	RequiredAcks int32 `json:"requiredAcks,omitempty" protobuf:"bytes,5,opt,name=requiredAcks"`
 	// Compress determines whether to compress message or not.
 	// Defaults to false.
 	// If set to true, compresses message using snappy compression.
@@ -495,7 +495,7 @@ type KafkaTrigger struct {
 	// FlushFrequency refers to the frequency in milliseconds to flush batches.
 	// Defaults to 500 milliseconds.
 	// +optional
-	FlushFrequency int `json:"flushFrequency,omitempty" protobuf:"bytes,7,opt,name=flushFrequency"`
+	FlushFrequency int32 `json:"flushFrequency,omitempty" protobuf:"bytes,7,opt,name=flushFrequency"`
 	// TLS configuration for the Kafka producer.
 	// +optional
 	TLS *TLSConfig `json:"tls,omitempty" protobuf:"bytes,8,opt,name=tls"`
@@ -677,7 +677,7 @@ type StatusPolicy struct {
 	// Allow refers to the list of allowed response statuses. If the response status of the the trigger is within the list,
 	// the trigger will marked as successful else it will result in trigger failure.
 	// +listType=allowedStatuses
-	Allow []int `json:"allow" protobuf:"bytes,1,name=allow"`
+	Allow []int32 `json:"allow" protobuf:"bytes,1,name=allow"`
 }
 
 // Backoff for an operation
@@ -689,7 +689,7 @@ type Backoff struct {
 	// The amount of jitter applied each iteration
 	Jitter float64 `json:"jitter" protobuf:"bytes,3,opt,name=jitter"`
 	// Exit with error after this many steps
-	Steps int `json:"steps" protobuf:"bytes,4,opt,name=steps"`
+	Steps int32 `json:"steps" protobuf:"bytes,4,opt,name=steps"`
 }
 
 // SensorResources holds the metadata of the resources created for the sensor
